refactor: relocate flat-schema parsing into _impl/schema_0
Move homie.py, accumulator.py, and field_metadata.py from mqtt/ into a
new _impl/schema_0/ package, and split the flat-schema Homie constants
(topic formats, TYPE_*, lugs direction, circuit id helpers) out of
mqtt/const.py into _impl/schema_0/const.py. mqtt/const.py keeps only
transport-level constants (HOMIE_STATE_*, MQTT_* connection settings).

Pure relocation: only import statements changed in the moved files.
Public API is unchanged (HomiePropertyAccumulator re-exported from the
new location; __all__ verified identical before/after).
