+++ src/span_panel_api/_impl/__init__.py
@@ -1,0 +1,1 @@
+"""Internal implementation packages. Not public API."""
