+++ src/span_panel_api/_impl/schema_0/__init__.py
@@ -1,0 +1,1 @@
+"""Flat-schema (Homie v5) parsing implementation. Not public API."""
