+++ src/span_panel_api/_impl/schema_0/consumer.py
@@ -12,9 +12,8 @@ import logging
 import time
 from typing import ClassVar
 
-from ..models import SpanBatterySnapshot, SpanCircuitSnapshot, SpanEvseSnapshot, SpanPanelSnapshot, SpanPVSnapshot
-from .accumulator import HomiePropertyAccumulator
-from .const import (
+from span_panel_api._impl.schema_0.accumulator import HomiePropertyAccumulator
+from span_panel_api._impl.schema_0.const import (
     LUGS_DOWNSTREAM,
     LUGS_UPSTREAM,
     TYPE_BESS,
@@ -27,6 +26,13 @@ from .const import (
     TYPE_POWER_FLOWS,
     TYPE_PV,
     normalize_circuit_id,
+)
+from span_panel_api.models import (
+    SpanBatterySnapshot,
+    SpanCircuitSnapshot,
+    SpanEvseSnapshot,
+    SpanPanelSnapshot,
+    SpanPVSnapshot,
 )
 
 _LOGGER = logging.getLogger(__name__)
