+++ src/span_panel_api/mqtt/__init__.py
@@ -1,10 +1,11 @@
 """SPAN Panel MQTT/Homie transport."""
 
-from .accumulator import HomieLifecycle, HomiePropertyAccumulator
+from span_panel_api._impl.schema_0.accumulator import HomieLifecycle, HomiePropertyAccumulator
+from span_panel_api._impl.schema_0.consumer import HomieDeviceConsumer
+
 from .async_client import AsyncMQTTClient
 from .client import SpanMqttClient
 from .connection import AsyncMqttBridge
-from .homie import HomieDeviceConsumer
 from .models import MqttClientConfig
 
 __all__ = [
