+++ src/span_panel_api/mqtt/client.py
@@ -13,15 +13,17 @@ import contextlib
 import logging
 import time
 
+from span_panel_api._impl.schema_0.accumulator import HomiePropertyAccumulator
+from span_panel_api._impl.schema_0.const import PROPERTY_SET_TOPIC_FMT, TYPE_CORE, WILDCARD_TOPIC_FMT
+from span_panel_api._impl.schema_0.consumer import HomieDeviceConsumer
+from span_panel_api._impl.schema_0.field_metadata import build_field_metadata, log_schema_drift
+
 from ..auth import get_homie_schema
 from ..exceptions import SpanPanelConnectionError, SpanPanelServerError, SpanPanelStaleDataError
 from ..models import FieldMetadata, HomieSchemaTypes, SpanPanelSnapshot
 from ..protocol import PanelCapability
-from .accumulator import HomiePropertyAccumulator
 from .connection import AsyncMqttBridge
-from .const import MQTT_READY_TIMEOUT_S, PROPERTY_SET_TOPIC_FMT, TYPE_CORE, WILDCARD_TOPIC_FMT
-from .field_metadata import build_field_metadata, log_schema_drift
-from .homie import HomieDeviceConsumer
+from .const import MQTT_READY_TIMEOUT_S
 from .models import MqttClientConfig
 
 _LOGGER = logging.getLogger(__name__)
